fundacionpet/requirements.txt
pillow
django-admin-interface
django-crispy-forms